add MatDialog,solve route issue, & give button in user_profile - AH

+++ client/src/app/pages/admin/admin.component.ts
@@ -2,6 +2,8 @@ import { Component } from '@angular/core';
 import { NavbarComponent } from '../../components/navbar/navbar.component';
 import { UsersService } from '../../shared/services/users.service';
 import { AsyncPipe, CommonModule } from '@angular/common';
+import { DeleteDialogComponent } from '../order/delete-dialog/delete-dialog.component';
+import { MatDialog } from '@angular/material/dialog';
 
 @Component({
   selector: 'app-admin',
@@ -13,7 +15,7 @@ import { AsyncPipe, CommonModule } from '@angular/common';
 export class AdminComponent {
   users:any[]=[];
   errMsg!:string;
-  constructor(private _UserService: UsersService){}
+  constructor(private _UserService: UsersService,private dialog: MatDialog){}
 
   ngOnInit():void {
     this._UserService.getAllUsers().subscribe({
@@ -22,11 +24,21 @@ export class AdminComponent {
   }
 
   deleteUser(id:number){
-    this._UserService.deleteUser(id).subscribe((res)=>console.log(res));
-    const index = this.users.findIndex(user => user._id === id);
-    if (index !== -1) {
-      this.users.splice(index, 1);
-    }
+    const dialogRef = this.dialog.open(DeleteDialogComponent);
+
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this._UserService.deleteUser(id).subscribe((res)=>console.log(res));
+        const index = this.users.findIndex(user => user._id === id);
+        if (index !== -1) {
+          this.users.splice(index, 1);
+        }
+      } else {
+        console.log('Delete cancelled');
+      }
+    });
+
+   
   }
 
 }
